Solution to INFRA-75
Helen raised INFRA-75 concerning the retention of an individual file in
the aspace-oai-s3-stage bucket. The current tf-mod-s3-iam terraform
module does not support the ability to configure tag-based or index-
based retention/expiration policies. The temporary fix is to just
disable the retention rule completely for this S3 bucket until we can
upgrade the tf-mod-s3-iam module to allow for granular retention
policies.

There is only one change in this commit, but there is a larger impact
when this is deployed because it will create new credentials for the
three IAM users.