Initial ETD Bucket Configuration
ETD-83 in Jira requries an S3 bucket for storing ingested theses as
part of the new ETD process. This is the initial commit for that task.
This requires an S3 bucket, and IAM user, CORS rules, and
retention/expiration rules for the bucket.

See the README for details. This creates the bucket, IAM user,
IAM rules, CORS rules, and content retention rules.